requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
babel==2.17.0
    # via -r requirements.in
certifi==2025.6.15
    # via requests
cffi==1.17.1
    # via vosk
charset-normalizer==3.4.2
    # via requests
colorama==0.4.6
    # via -r requirements.in
evdev==1.9.2
    # via pynput
idna==3.10
    # via requests
pyaudio==0.2.14
    # via -r requirements.in
pycparser==2.22
    # via cffi
pynput==1.7.6
    # via -r requirements.in
pyperclip==1.9.0
    # via -r requirements.in
python-xlib==0.33
    # via pynput
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.4
    # via vosk
six==1.17.0
    # via
    #   pynput
    #   python-xlib
srt==3.5.3
    # via vosk
tqdm==4.67.1
    # via vosk
urllib3==2.5.0
    # via requests
vosk==0.3.45
    # via -r requirements.in
websockets==15.0.1
    # via vosk
